checklist_system/requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
qrcode[pil]==7.4.2
python-dotenv==1.0.1
email-validator==2.1.1